Fixed some animations issues and add standing roll
I've fixed the main bug I had with the animation which was (ofc) caused by inattention, Roll NW was RUN NW so yeah, that was it. I had to switch Roll NW with NE to realize that the problem was not in the script but in the animator controller.
I've also had a was for the player the stay in the Idle state facing the last direction he was moving. Using this I had a way for the player to roll in that direction if he is not moving and pressing the jump button.
I also make is so that the sprint animation stops playing when the player isnt moving.
I also add a way for the isMoving bool to be set back to false... :)